refactor: move IndexedDB implementation to core storage layer
Move the actual IndexedDB implementation from `src/utils/indexedDB.ts`
into `src/core/storage/backends/indexedDB.ts` to fix architectural
violation where core storage infrastructure imported from app-level
utilities.

This change:
- Eliminates upward dependency (core -> app-level)
- Places storage infrastructure in the proper core layer
- Maintains proper layered architecture (core <- shared <- features)

Also relocates tests to match the new module location.
