+++ src/test/storage-backend.test.ts
@@ -6,11 +6,11 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import * as backend from '@/core/storage/backend';
 import * as localStorage from '@/core/storage/backends/localStorage';
-import * as indexedDB from '@/utils/indexedDB';
+import * as indexedDB from '@/core/storage/backends/indexedDB';
 import type { Layout } from '@/core/types';
 
 // Mock IndexedDB module
-vi.mock('../utils/indexedDB', () => ({
+vi.mock('../core/storage/backends/indexedDB', () => ({
   saveLayout: vi.fn(),
   loadLayout: vi.fn(),
   deleteLayout: vi.fn(),
